put in square literal names instead of numbers bench 4884490

+++ src/perft.rs
@@ -6,7 +6,7 @@ pub fn perft(board: &Board, depth: i32) -> usize {
     let start = Instant::now();
     let count = non_bulk_perft::<true>(board, depth);
     let elapsed = start.elapsed().as_secs_f64();
-    println!("{count} nodes in {elapsed} secs = {} nps", count as f64 / elapsed);
+    println!("{count} nodes in {elapsed} secs = {} nps", (count as f64 / elapsed) as u64);
     count
 }
